requirement.txt
streamlit
pandas
folium
requests
streamlit-folium
python-dotenv